add Crush 2

+++ internal/repository/cart_repository.go
@@ -2,7 +2,9 @@ package repository
 
 import (
 	"context"
+	"errors"
 	"fmt"
+
 	"github.com/google/uuid"
 	"github.com/nikolayk812/sqlcpp-demo/internal/db"
 	"github.com/nikolayk812/sqlcpp-demo/internal/domain"
@@ -10,24 +12,31 @@ import (
 	"golang.org/x/text/currency"
 )
 
+var (
+	ErrOwnerIDEmpty = errors.New("ownerID is empty")
+)
+
 type cartRepository struct {
-	q    *db.Queries
-	dbtx db.DBTX
+	q *db.Queries
 }
 
+// NewCart creates a new CartRepository with the given dbtx (pgx.Tx or pgxpool.Pool).
 func NewCart(dbtx db.DBTX) (port.CartRepository, error) {
 	if dbtx == nil {
 		return nil, fmt.Errorf("dbtx is nil")
 	}
 
 	return &cartRepository{
-		q:    db.New(dbtx),
-		dbtx: dbtx,
+		q: db.New(dbtx),
 	}, nil
 }
 
 func (r *cartRepository) GetCart(ctx context.Context, ownerID string) (domain.Cart, error) {
 	var cart domain.Cart
+
+	if ownerID == "" {
+		return cart, ErrOwnerIDEmpty
+	}
 
 	rows, err := r.q.GetCart(ctx, ownerID)
 	if err != nil {
@@ -38,9 +47,9 @@ func (r *cartRepository) GetCart(ctx context.Context, ownerID string) (domain.Ca
 	cart.Items = make([]domain.CartItem, 0, len(rows))
 
 	for _, row := range rows {
-		item, err := mapGetCartRowToDomainCartItem(row)
+		item, err := mapGetCartRowToCartItem(row)
 		if err != nil {
-			return cart, fmt.Errorf("mapGetCartRowToDomainCartItem: %w", err)
+			return cart, fmt.Errorf("mapGetCartRowToCartItem: %w", err)
 		}
 		cart.Items = append(cart.Items, item)
 	}
@@ -49,6 +58,10 @@ func (r *cartRepository) GetCart(ctx context.Context, ownerID string) (domain.Ca
 }
 
 func (r *cartRepository) AddItem(ctx context.Context, ownerID string, item domain.CartItem) error {
+	if ownerID == "" {
+		return ErrOwnerIDEmpty
+	}
+
 	err := r.q.AddItem(ctx, db.AddItemParams{
 		OwnerID:       ownerID,
 		ProductID:     item.ProductID,
@@ -63,6 +76,10 @@ func (r *cartRepository) AddItem(ctx context.Context, ownerID string, item domai
 }
 
 func (r *cartRepository) DeleteItem(ctx context.Context, ownerID string, productID uuid.UUID) (bool, error) {
+	if ownerID == "" {
+		return false, ErrOwnerIDEmpty
+	}
+
 	rowsAffected, err := r.q.DeleteItem(ctx, db.DeleteItemParams{
 		OwnerID:   ownerID,
 		ProductID: productID,
@@ -74,7 +91,8 @@ func (r *cartRepository) DeleteItem(ctx context.Context, ownerID string, product
 	return rowsAffected > 0, nil
 }
 
-func mapGetCartRowToDomainCartItem(row db.GetCartRow) (domain.CartItem, error) {
+// mapGetCartRowToCartItem maps a GetCartRow to a domain.CartItem
+func mapGetCartRowToCartItem(row db.GetCartRow) (domain.CartItem, error) {
 	parsedCurrency, err := currency.ParseISO(row.PriceCurrency)
 	if err != nil {
 		return domain.CartItem{}, fmt.Errorf("currency[%s] is not valid: %w", row.PriceCurrency, err)
